_DEMO_cocktails/requirements.txt
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
flatbuffers==2.0
gunicorn==20.1.0
install==1.3.5
Jinja2==3.1.2
numpy==1.19.5
openvino==2022.1.0
Pillow==9.1.1
protobuf==4.21.1
wget==3.2
WTForms==3.0.1